allow for distinct initial metrics from subsequent metrics (#11)
Sometimes, you get a dataset that you need to start `count(*)`, and then
you need to compare redactions by summing that metric.

for example, first you need to`count(*) as incidence` from vaccine-level
data to get the initial aggregation. Then to do latent checks to you
need to compare incidences; if you forget to do this and count(*) again
from that initial aggregation that gets cached, you'll accidentally
count your aggregated rows, which is undesirable as you are comparing
the wrong metric for redaction.